+++ pkg/store/store_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/go-redis/redis/v8"
+	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/config"
 	"github.com/mvisonneau/gitlab-ci-pipelines-exporter/pkg/schemas"
 	"github.com/stretchr/testify/assert"
 )
@@ -27,4 +28,27 @@ func TestNewRedisStore(t *testing.T) {
 	}
 
 	assert.Equal(t, expectedValue, NewRedisStore(redisClient))
+}
+
+func TestNew(t *testing.T) {
+	localStore := New(nil, config.Projects{})
+	assert.IsType(t, &Local{}, localStore)
+
+	redisClient := redis.NewClient(&redis.Options{})
+	redisStore := New(redisClient, config.Projects{})
+	assert.IsType(t, &Redis{}, redisStore)
+
+	localStore = New(nil, config.Projects{
+		{
+			Name: "foo",
+		},
+		{
+			Name: "foo",
+		},
+		{
+			Name: "bar",
+		},
+	})
+	count, _ := localStore.ProjectsCount()
+	assert.Equal(t, int64(2), count)
 }